+++ src/model.rs
@@ -41,4 +41,5 @@ pub type ModelResult<T, E = ModelError> = std::result::Result<T, E>;
 #[async_trait]
 pub trait Authenticable: Clone {
     async fn find_by_api_key(db: &DatabaseConnection, api_key: &str) -> ModelResult<Self>;
+    async fn find_by_claims_key(db: &DatabaseConnection, claims_key: &str) -> ModelResult<Self>;
 }